requirements.txt
alembic==1.11.1
anyio==3.7.0
attrs==23.1.0
bcrypt==4.0.1
certifi==2023.5.7
cffi==1.15.1
chardet==4.0.0
click==8.1.3
dnspython==2.3.0
docstring-parser==0.15
email-validator==1.3.1
Faker==18.10.1
fastapi==0.96.0
FastAPI-SQLAlchemy==0.2.1
h11==0.14.0
idna==2.10
importlib-metadata==6.6.0
iniconfig==2.0.0
Mako==1.2.4
MarkupSafe==2.1.3
numpy==1.24.3
packaging==23.1
pandas==2.0.2
passlib==1.7.4
pluggy==1.0.0
psycopg2-binary==2.9.6
py==1.11.0
pycparser==2.21
pydantic==1.10.8
PyJWT==2.7.0
pyparsing==3.0.9
pytest==7.3.1
python-dateutil==2.8.2
python-dotenv==0.15.0
python-editor==1.0.4
python-multipart==0.0.6
pytz==2023.3
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.15
requests==2.31.0
starlette==0.27.0
text-unidecode==1.3
toml==0.10.2
typing_extensions==4.6.3
urllib3==1.26.16
uvicorn==0.22.0
zipp==3.15.0
boto3==1.26.147
b2sdk~=1.21.0
botocore==1.29.147
websockets~=11.0.3
pyfcm==1.5.4
aiofiles==23.1.0
celery==5.2.7
Jinja2==3.1.2
redis==4.5.5
httpx==0.24.1
flower==1.2.0
fastapi_mail==1.2.8